Add file merging function (heap sort)
The program uses the concept of heap sort to merge files (created
in split_file_and_sort), and it writes the data to the output
file.

How it works:
1. Open the sorted files and store the file pointers in a file
pointer array.

2. Read the first integers in each file, store the integers and
the corresponding file indices in the node_t array.

3. Create a heap and use heap sort to arrange the integers in
ascending order.

4. Write value of the smallest node to the output file and mark
the file index (where the smallest node came from) as the current
file index.

5. Read the next integer from the current file. If EOF is
reached, go to step 6.1. Otherwise, go to step 6.2.

6.1. Swap the smallest node with the last node in the heap.
Decrease size of the heap by 1. If the size is zero, go to step
9, otherwise, go to step 7.

6.2. Replace the smallest node in the heap with the value from
step 5.

7. Update the heap by using min-heapify.

8. Write value of the updated smallest node to the output file
and update the current file index if necessary. Repeat step 5.

9. Done!