Fix error of opening too many files at once
Originally, the sorting program will crash when it tries to open
more than 1021 files at the same time. Therefore, I set MAX_FILES
to 1000 to prevent the program from crashing. After the input
file is splitted, there might be more than 1000 files, so the
program will merge 1000 files each time until the number of files
become one, which means that the merging process is finished, and
the external sorting is done.